@@ -16,10 +16,10 @@ dependencies:
     path: '../settings'
 
   # BLoC - https://pub.dev/packages/flutter_bloc
-  flutter_bloc: ^8.1.4
+  flutter_bloc: ^8.1.6
 
   # Google Maps - https://pub.dev/packages/google_maps_flutter
-  google_maps_flutter: ^2.6.0
+  google_maps_flutter: ^2.9.0
   google_maps_flutter_android:
   google_maps_flutter_platform_interface:
 
@@ -27,30 +27,28 @@ dependencies:
   apple_maps_flutter: ^1.3.0
 
   # MapLibre - https://github.com/maplibre/flutter-maplibre-gl
-  maplibre_gl:
-    git:
-      url: https://github.com/fabi755/flutter-maplibre-gl.git
-      ref: feature/gradle-namespace
+  maplibre_gl: ^0.20.0
+#    git:
+#      url: https://github.com/fabi755/flutter-maplibre-gl.git
+#      ref: feature/gradle-namespace
 
   # Multiple Result - https://pub.dev/packages/multiple_result
   multiple_result: ^5.1.0
 
   # Rx Dart - https://pub.dev/packages/rxdart
-  rxdart: ^0.27.7
+  rxdart: ^0.28.0
 
   # Shared Preferences - https://pub.dev/packages/shared_preferences
-  shared_preferences: ^2.2.2
+  shared_preferences: ^2.3.2
 
-dependency_overrides:
-  # Force `package_info_plus` using newest `web` version (see: https://github.com/fluttercommunity/plus_plugins/issues/2653)
-  web: ^0.5.0
-
-  # Force `flutter_localizations` using newest `intl`
-  intl: ^0.19.0
+#dependency_overrides:
+#  # Force `package_info_plus` using newest `web` version (see: https://github.com/fluttercommunity/plus_plugins/issues/2653)
+#  web: ^0.5.0
+#
+#  # Force `flutter_localizations` using newest `intl`
+#  intl: ^0.19.0
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  flutter_lints: ^3.0.1
+  flutter_lints: ^5.0.0
 
 flutter:
