@@ -22,19 +22,19 @@ dependencies:
     path: '../station'
 
   # BLoC - https://pub.dev/packages/flutter_bloc
-  flutter_bloc: ^8.1.4
+  flutter_bloc: ^8.1.6
 
   # Shared Preferences - https://pub.dev/packages/shared_preferences
-  shared_preferences: ^2.2.2
+  shared_preferences: ^2.3.2
 
   # Font Awesome - https://pub.dev/packages/font_awesome_flutter
-  font_awesome_flutter: ^10.7.0
+  font_awesome_flutter: ^10.8.0
 
   # URL Launcher - https://pub.dev/packages/url_launcher
-  url_launcher: ^6.2.5
+  url_launcher: ^6.3.1
 
   # Map launcher - https://pub.dev/packages/map_launcher
-  map_launcher: ^3.2.0
+  map_launcher: ^3.5.0
 
   # Collection - https://pub.dev/packages/collection
   collection: ^1.18.0
@@ -46,24 +46,22 @@ dependencies:
   multiple_result: ^5.1.0
 
   # Rx Dart - https://pub.dev/packages/rxdart
-  rxdart: ^0.27.7
+  rxdart: ^0.28.0
 
   # Easy Localization - https://pub.dev/packages/easy_localization
-  easy_localization: ^3.0.5
+  easy_localization: ^3.0.7
 
     # Share Plus - https://pub.dev/packages/share_plus
-  share_plus: 8.0.3
+  share_plus: 10.1.1
 
-dependency_overrides:
-  # Force `package_info_plus` using newest `web` version (see: https://github.com/fluttercommunity/plus_plugins/issues/2653)
-  web: ^0.5.0
-
-  # Force `flutter_localizations` using newest `intl`
-  intl: ^0.19.0
+#dependency_overrides:
+#  # Force `package_info_plus` using newest `web` version (see: https://github.com/fluttercommunity/plus_plugins/issues/2653)
+#  web: ^0.5.0
+#
+#  # Force `flutter_localizations` using newest `intl`
+#  intl: ^0.19.0
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  flutter_lints: ^3.0.1
+  flutter_lints: ^5.0.0
 
 flutter:
