@@ -25,25 +25,25 @@ dependencies:
     path: '../report'
 
   # BLoC - https://pub.dev/packages/flutter_bloc
-  flutter_bloc: ^8.1.4
+  flutter_bloc: ^8.1.6
 
   # Rx Dart - https://pub.dev/packages/rxdart
-  rxdart: ^0.27.7
+  rxdart: ^0.28.0
 
   # HTTP - https://pub.dev/packages/http
-  http: ^1.2.1
+  http: ^1.2.2
 
   # Google Maps - https://pub.dev/packages/google_maps_flutter
-  google_maps_flutter: ^2.6.0
+  google_maps_flutter: ^2.9.0
 
   # Shared Preferences - https://pub.dev/packages/shared_preferences
-  shared_preferences: ^2.2.2
+  shared_preferences: ^2.3.2
 
   # Location - https://pub.dev/packages/geolocator
-  geolocator: ^11.0.0
+  geolocator: ^13.0.1
 
   # In-App review - https://pub.dev/packages/in_app_review
-  in_app_review: ^2.0.8
+  in_app_review: ^2.0.9
 
   # Multiple Result - https://pub.dev/packages/multiple_result
   multiple_result: ^5.1.0
@@ -55,21 +55,19 @@ dependencies:
   collection: ^1.18.0
 
   # URL Launcher - https://pub.dev/packages/url_launcher
-  url_launcher: ^6.2.5
+  url_launcher: ^6.3.1
 
   # Easy Localization - https://pub.dev/packages/easy_localization
-  easy_localization: ^3.0.5
+  easy_localization: ^3.0.7
 
-dependency_overrides:
-  # Force `package_info_plus` using newest `web` version (see: https://github.com/fluttercommunity/plus_plugins/issues/2653)
-  web: ^0.5.0
-
-  # Force `flutter_localizations` using newest `intl`
-  intl: ^0.19.0
+#dependency_overrides:
+#  # Force `package_info_plus` using newest `web` version (see: https://github.com/fluttercommunity/plus_plugins/issues/2653)
+#  web: ^0.5.0
+#
+#  # Force `flutter_localizations` using newest `intl`
+#  intl: ^0.19.0
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  flutter_lints: ^3.0.1
+  flutter_lints: ^5.0.0
 
 flutter:
